Preparation for merge with master
Consolidation and housekeeping: removed files that are no longer in use / were for development purposes.
